requirements.txt
colorama
pwntools
tqdm